+++ src/content/index.tsx
@@ -3,8 +3,7 @@ import ReactDOM from "react-dom/client";
 import HighlightBrackets from "./HighlightBrackets";
 import HighlightParagraph from "./HightlightParagraph";
 import HoverTooltip from "./HoverTooltip";
-// import HoverPopup from "./HoverPopup";
-import LawPopover from "./LawPopover";
+import ExternalLinkPopover from "./external-link-popover/ExternalLinkPopover";
 
 function initialize() {
   const rootDiv = document.createElement("div");
@@ -14,11 +13,10 @@ function initialize() {
   const root = ReactDOM.createRoot(rootDiv);
   root.render(
     <>
-      <LawPopover />
+      <ExternalLinkPopover />
       <HighlightBrackets />
       <HighlightParagraph />
       <HoverTooltip />
-      {/* <HoverPopup /> */}
     </>
   );
 }
