+++ src/utils/utils.ts
@@ -27,123 +27,3 @@ export async function querySelectorAllWithDelay(
     await new Promise((resolve) => setTimeout(resolve, delay));
   }
 }
-
-/**
- * law_full_text（JSON形式）をテキスト化するための再帰関数
- */
-function parseLawNode(node: any, depth: number = 0): string {
-  // 1. もし node が文字列なら、そのまま返す
-  if (typeof node === "string") {
-    return node;
-  }
-
-  // 2. node がオブジェクトの場合、tag や attr, children を取得
-  const { tag, attr, children = [] } = node;
-
-  // タグに応じて見出しなどを軽く整形してみる (必要に応じて拡張)
-  let prefix = "";
-  let suffix = "";
-
-  switch (tag) {
-    case "LawTitle":
-    case "PartTitle":
-    case "ChapterTitle":
-    case "SectionTitle":
-    case "SubsectionTitle":
-    case "ArticleTitle":
-      // 見出し系は前後に改行を挟む
-      prefix = "\n\n";
-      suffix = "\n";
-      break;
-    case "Article":
-      // Article は記事番号を表示させたりもできる (attr.Num)
-      // ここでは単純に前後に改行だけ
-      prefix = "\n";
-      suffix = "\n";
-      break;
-    case "Paragraph":
-      // 段落ごとに改行
-      prefix = "\n";
-      suffix = "\n";
-      break;
-    case "Item":
-      // 号などを表示
-      prefix = "\n  ・";
-      break;
-    case "Sentence":
-      // 文章の続きなのであまり区切らないが、
-      // 例として改行入れたい場合は suffix = "\n"; にしても良い
-      break;
-    default:
-      break;
-  }
-
-  // 3. children を再帰的にパースし、連結
-  let content = children
-    .map((child: any) => parseLawNode(child, depth + 1))
-    .join("");
-
-  // 4. 必要に応じて、タグが "ArticleTitle" のときに記事番号を付けるなど拡張可能
-  //   例: if (tag === "ArticleTitle") { content = `【${content}】`; }
-
-  return prefix + content + suffix;
-}
-
-/**
- * law_full_text のオブジェクト（JSON）をまとめてテキスト化
- */
-export function parseLawFullText(lawFullText: any): string {
-  return parseLawNode(lawFullText);
-}
-
-/**
- * フラグメント文字列(#以降) → 法令APIの elm (例: MainProvision-Article_36) へ変換
- * 例:
- *   "#Mp-Pa_1-Ch_3-Se_1-At_36-Pr_1"
- *   => "MainProvision-Part_1-Chapter_3-Section_1-Article_36-Paragraph_1"
- */
-export function fragmentToElm(fragment: string): string | null {
-  // 先頭に # があれば除去
-  const frag = fragment.replace(/^#/, "");
-
-  const mapping: Record<string, string> = {
-    Mp: "MainProvision",
-    Sp: "SupplProvision",
-    Pa: "Part",
-    Ch: "Chapter",
-    Se: "Section",
-    Su: "Subsection",
-    Di: "Division",
-    At: "Article",
-    Pr: "Paragraph",
-  };
-
-  const parts = frag.split("-");
-  if (!parts.length) return null;
-
-  // MainProvision or SupplProvision の処理
-  const root = parts[0];
-  let resultArray: string[] = [];
-  if (root === "Mp") {
-    resultArray.push("MainProvision");
-  } else if (root === "Sp") {
-    resultArray.push("SupplProvision");
-  } else {
-    console.warn("Unknown fragment root:", root);
-    return null;
-  }
-
-  for (let i = 1; i < parts.length; i++) {
-    // segment例: "Pa_1"
-    const seg = parts[i];
-    const [code, num] = seg.split("_");
-    const mapped = mapping[code];
-    if (!mapped) {
-      console.warn("Unknown segment code:", code);
-      return null;
-    }
-    resultArray.push(`${mapped}_${num}`);
-  }
-
-  return resultArray.join("-");
-}
